Encapsulate status text builder
